+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 ### Changed
+* Changed to update job status to error when celery exception error
+  Contributed by @hinashi
 
 * (New-UI) Show a shorten name if entity name is too long
   at Entity list view.